requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
newsapi-python==0.2.6
PyJWT==2.0.0
pytz==2020.5
requests==2.25.1
six==1.15.0
twilio==6.50.1
urllib3==1.26.2
